+++ pkg/admin/services/abstract_service.go
@@ -1,7 +1,5 @@
 package services
 
-import "github.com/apache/dubbo-admin/pkg/admin/registry/config"
-
 //func a() {
 //	extension.S
 //}
@@ -13,9 +11,9 @@ import "github.com/apache/dubbo-admin/pkg/admin/registry/config"
 //var metaDataCollector MetaDataCollector
 //var interfaceRegistryCache InterfaceRegistryCache = NewInterfaceRegistryCache()
 
-type AbstractService struct {
-	dynamicConfiguration config.GovernanceConfiguration
-}
+//type AbstractService struct {
+//	dynamicConfiguration config.GovernanceConfiguration
+//}
 
 //func (service *AbstractService) GetInterfaceRegistryCache() sync.Map {
 //	return interfaceRegistryCache.GetRegistryCache()
